+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to aws organization formation will be documented in this file.
 
+**not released**
+- Added --stack-trace flag to print stack traces for errors. stacktraces are now hidden by default.
+
 **version 0.0.68**
 - Added support for passing !Ref account in update-stacks parameters (tasks file).
 - Added validation on the tasks in tasks file: top level file must have exactly 1 update-organization task